examples(incident-triage): collapse to a single agent; drop the larger full demo
Removes the larger ``examples/incident_triage_agent/`` (SOP-runbook variant
with a separate ``fixtures/`` dir) and renames ``examples/incident_triage_simple/``
into its place using ``git mv`` so the simpler customer-facing agent now
owns the canonical ``incident_triage_agent`` path.

Files removed (full-demo variant):
- examples/incident_triage_agent/SOP.md
- examples/incident_triage_agent/fixtures/* (the JSON-per-incident layout)

Files preserved (simple-agent variant, renamed in place):
- examples/incident_triage_agent/__init__.py
- examples/incident_triage_agent/.env.example
- examples/incident_triage_agent/agent.py
- examples/incident_triage_agent/eval_config.yaml
- examples/incident_triage_agent/README.md
- examples/incident_triage_agent/fixtures.json

Cross-references updated:
- examples/README.md: one incident_triage_agent entry, simple-agent description.
- eval_config.yaml: callable path examples.incident_triage_simple.agent -> examples.incident_triage_agent.agent.
- agent.py: internal imports updated if any.
- tests/test_incident_triage_smoke.py: paths updated; full-demo-only assertions removed.

Verification:
- pytest tests/test_incident_triage_smoke.py -> passes
- git grep 'incident_triage_simple' -> zero results
- git grep 'incident-triage-simple' -> zero results
