+++ examples/incident_triage_agent/.env.example
@@ -1,16 +1,7 @@
-# Incident Triage Agent — local environment overrides
+# Incident Triage (Simple) — local environment.
 #
-# This file layers on top of the repo-root .env (LiteLLM provider creds shared
-# with all demos). It only needs to override demo-specific knobs.
-#
-# In the repo-root .env, you must define:
-#   AZURE_API_KEY
-#   AZURE_API_BASE       (e.g. https://<resource>.openai.azure.com)
-#   AZURE_API_VERSION    (e.g. 2024-10-21)
+# Inherits from the repo-root .env (LiteLLM provider creds shared across demos).
+# In the repo-root .env you must define: AZURE_API_KEY, AZURE_API_BASE, AZURE_API_VERSION.
 
-# Model used by the agent's reasoning loop (LiteLLM provider/model spelling).
 INCIDENT_TRIAGE_MODEL=azure/gpt-5.4-mini
-
-# Phoenix project name (optional — enables traces under a dedicated project
-# in the local Phoenix UI at http://localhost:6006).
 PHOENIX_PROJECT_NAME=incident-triage-agent
