+++ examples/incident_triage_agent/fixtures.json
