--- examples/incident_triage_simple/.env.example
@@ -1,7 +1,0 @@
-# Incident Triage (Simple) — local environment.
-#
-# Inherits from the repo-root .env (LiteLLM provider creds shared across demos).
-# In the repo-root .env you must define: AZURE_API_KEY, AZURE_API_BASE, AZURE_API_VERSION.
-
-INCIDENT_TRIAGE_MODEL=azure/gpt-5.4-mini
-PHOENIX_PROJECT_NAME=incident-triage-simple
