--- examples/incident_triage_simple/__init__.py
@@ -1,2 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT License.
